requirements.txt
fastapi>=0.100.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.23.0
pyannote.audio
openai-whisper
vosk
torch
numpy
pydub
librosa
datasets    
transformers
speechbrain
requests>=2.31.0
aiofiles>=23.0.0
python-multipart>=0.0.6
# LLM Provider APIs
openai>=1.0.0
google-cloud-aiplatform
anthropic
# Database and Storage
sqlalchemy
psycopg2-binary
chromadb
# Blockchain and Web3
web3
eth-account
# Development dependencies
pytest
pytest-asyncio
httpx
pydantic>=2.0.0